Port diff patch tests
