--- tests/jsdom/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "jsdom-tests"
-version = "0.1.0"
-authors = ["Chinedu Francis Nwafili <frankie.nwafili@gmail.com>"]
-publish = false
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-virtual-dom-rs = { path = "../../virtual-dom-rs" }
-
-[dependencies.js-sys]
-version = "0.2"
-
-[dependencies.wasm-bindgen]
-version = "0.2"
-features = ["default", "nightly"]
-
-[dependencies.web-sys]
-version = "0.3"
-features = [
-    "console",
-    "Window",
-    "Document",
-    "HtmlElement",
-]
