--- tests/jsdom/README.md
@@ -1,3 +1,0 @@
-# jsdom tests
-
-Test our virtual-dom-rs library in a DOM using [jsdom](https://github.com/jsdom/jsdom), Node.js and WebAssembly.
